DarkTracks.jl

DarkTracks.jl

Track a bright spot in a dim video. The initial location of the spot is assumed to be somewhere in the middle of the video-frame. The video should be more or less uniformly dark. The bright spot doesn't need to be visible all of the time, but should be fairly brighter than the background when visible.

Prerequisites

You have to have ffmpeg installed (this package also utilizes ffprobe, which comes bundled with ffmpeg).

Table of contents

DarkTracks.jl

Prerequisites

Table of contents

Functions

Index

Functions

DarkTracks.main — Function.

main(videofile; start_frame = 1, stop_frame = lastframe(videofile),  spatialscale = 10, temporalscale = 10)

Track a bright spot in a dim video file, videofile. Optionally give a frame number to start at, start_frame, and stop, stop_frame. These default to the beginning and ending of the video file. spatialscale and temporalscale help subsample the video for faster processing. When these equal 1 then there is no sub-sampling and the video is processed in original spatial and temporal resolutions. For example, for scaling the frame by half its original size and sampling every second frame use spatialscale = 2, temporalscale = 2. A default value of 10 for both works well.

Examples

julia> main(videofile, stop_frame = 100, spatialscale = 10)
([1.32, 1.54, 1.76, 1.98], Tuple{Int64,Int64}[(510, 730), (520, 730), (520, 730), (520, 730)])

source

Index

DarkTracks.main