DarkTracks.jl

DarkTracks.jl

Track a bright spot in a dim video. The initial location of the spot is assumed to be somewhere in the middle of the video-frame. The video should be more or less uniformly dark. The bright spot doesn't need to be visible all of the time, but should be fairly brighter than the background when visible.

Prerequisites

You have to have ffmpeg installed (this package also utilizes ffprobe, which comes bundled with ffmpeg).

Installation

If you haven't already, install the current release of Julia -> you should be able to launch it (some icon on the Desktop or some such).

Start Julia -> a Julia-terminal popped up.

Copy:
using Pkg
Pkg.add(PackageSpec(url = "https://github.com/yakir12/DarkTracks.jl"))
and paste it in the newly opened Julia-terminal, press Enter.

(not necessary) To test the package, copy:
Pkg.test("DarkTracks")
and paste it in the Julia-terminal. Press enter to check if all the tests pass -> this may take some time (all tests should pass).

You can close the Julia-terminal after it's done running.

Quick start

Start Julia -> a Julia-terminal popped up.

Copy and paste this in the newly opened Julia-terminal:
using DarkTracks
seconds, coordinates = main(videofile)
where videofile is the path to the video file you want to process. Please read about the different options the main function can accept.

Output

Apart from the return variables of the main function, the function saves a tab-delimited file with the seconds and coordinates of the tracked spot. The file name is the same name as the video file with a csv file extension and is located in the same directory as the video file.

Functions

DarkTracks.main — Function.

seconds, coordinates = main(videofile; start_frame = 1, stop_frame = lastframe(videofile),  spatialscale = 10, temporalscale = 10)

Track a bright spot in a dim video file, videofile. Optionally give a frame number to start at, start_frame, and stop, stop_frame. These default to the beginning and ending of the video file. spatialscale and temporalscale help subsample the video for faster processing. When these equal 1 then there is no sub-sampling and the video is processed in original spatial and temporal resolutions. For example, for scaling the frame by half its original size and sampling every second frame use spatialscale = 2, temporalscale = 2. A default value of 10 for both works well.

Return variables

seconds is a Vector{Float64} of times in second for each of the coordinates (itself a Vector{Tuple{Int, Int}}).

Examples

julia> main(videofile, stop_frame = 100, spatialscale = 10)
([1.32, 1.54, 1.76, 1.98], Tuple{Int64,Int64}[(510, 730), (520, 730), (520, 730), (520, 730)])

source

Index

DarkTracks.main